Add environment variable for git clone --reference
git clone's --reference[-if-able] option uses a local checkout of a
repository to seed the objects that would be fetched during a clone.
On the same filesystem this will result in hardlinks, and across
filesystems it uses symlinks. Using symlinks is an issue if the original
repository is removed, it will break the second repository.
To ensure the operation is safe, '--dissociate' is added to git clone
to instead copy the objects. This is likely still faster than network.

The environment varible is WIT_WORKSPACE_REFERENCE which should be a
directory containing repositories you would like to use as a seed,
such as another wit workspace.
Any repositories that are missing are cloned from the network as before.